requirements.txt
pyserial
web.py
attrs
cattrs